input

+++ QuanLyChuyenBay_Console/QuanLyMayBay.cpp
@@ -93,34 +93,24 @@ void QuanLyMayBay::update_byposition(int i, MayBay * value)
 	DanhSach[i]->SoDong = value->SoDong;
 }
 
-bool QuanLyMayBay::isSoHieuHopLe(string str)
+std::string QuanLyMayBay::Build_SoHieuMayBay()
 {
-	return str.size() > 0 && str.size() <= 15;
+	string res = "MB";
+	int max = 0;
+
+	for (int i = 0; i < SoLuongMayBay; i++) {
+		int id = StringToInteger(DanhSach[i]->SoHieuMayBay.substr(2, 14));
+
+		if (id > max) max = id;
+	}
+
+	res = res + std::to_string(max + 1);
+	return res;
 }
 
 bool QuanLyMayBay::isLoaiHopLe(string str)
 {
 	return str.size() > 0 && str.size() <= 40;
-}
-
-bool QuanLyMayBay::isSoDayHopLe(string str)
-{
-	for (int i = 0; i < str.size(); i++) {
-		if (!(str[i] >= '0' && str[i] <= '9'))
-			return false;
-	}
-
-	return true;
-}
-
-bool QuanLyMayBay::isSoDongHopLe(string str)
-{
-	for (int i = 0; i < str.size(); i++) {
-		if (!(str[i] >= '0' && str[i] <= '9'))
-			return false;
-	}
-
-	return true;
 }
 
 MayBay * QuanLyMayBay::getBy_SoHieuMayBay(std::string sohieu)
@@ -133,13 +123,27 @@ MayBay * QuanLyMayBay::getBy_SoHieuMayBay(std::string sohieu)
 	return NULL;
 }
 
+int QuanLyMayBay::getSoLuongMayBay()
+{
+	return SoLuongMayBay;
+}
+
+std::vector<MayBay> QuanLyMayBay::getDanhSachMayBay()
+{
+	std::vector<MayBay> danhsach;
+	for (int i = 0; i < SoLuongMayBay; i++)
+		danhsach.push_back(*DanhSach[i]);
+
+	return danhsach;
+}
+
 int QuanLyMayBay::Menu()
 {
 	while (1) {
 		Clrscr();
 		ShowList();
 
-		GotoXY(0, 0); SetColor(colorGreen); cout << "QUAN LY MAY BAY";
+		GotoXY(0, 0); SetColor(colorHeader); cout << "QUAN LY MAY BAY";
 		GotoXY(5, 2); SetColor(colorCyan); cout << "1. Them may bay";
 		GotoXY(5, 3); SetColor(colorCyan); cout << "2. Hieu chinh may bay";
 		GotoXY(5, 4); SetColor(colorCyan); cout << "3. Xoa may bay";
@@ -153,39 +157,46 @@ int QuanLyMayBay::Menu()
 
 			SetCursorVisible(1);
 			ClrLine(8); GotoXY(0, 8); SetColor(colorWhite); cout << "-> ";
-			cin >> choose;
+			input->GetInput();
 
-			switch (choose)
-			{
-			case 1:
-				if (Add() == -1) {
-					rollBack = 1;
-					break;
-				}
-			case 2:
-				if (Modify() == -1) {
-					rollBack = 1;
-					break;
-				}
-			case 3:
-				if (Delete() == -1) {
-					rollBack = 1;
-					break;
-				}
-			case 4:
-				if (Export() == -1) {
-					rollBack = 1;
-					break;
-				}
-			case 5:
-				return -1;
-			default:
-				SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
-				break;
+			if (!input->isInteger()) {
+				ClrLine(9); GotoXY(0, 9); SetColor(colorRed); cout << "ERROR: Input khong hop le!";
 			}
+			else {
+				choose = StringToInteger(input->GetResult());
 
-			if (rollBack == 1)
-				break;
+				switch (choose)
+				{
+				case 1:
+					if (Add() == -1) {
+						rollBack = 1;
+						break;
+					}
+				case 2:
+					if (Modify() == -1) {
+						rollBack = 1;
+						break;
+					}
+				case 3:
+					if (Delete() == -1) {
+						rollBack = 1;
+						break;
+					}
+				case 4:
+					if (Export() == -1) {
+						rollBack = 1;
+						break;
+					}
+				case 5:
+					return -1;
+				default:
+					ClrLine(9); GotoXY(0, 9); SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
+					break;
+				}
+
+				if (rollBack == 1)
+					break;
+			}		
 		}
 	}
 }
@@ -199,7 +210,7 @@ void QuanLyMayBay::ShowList()
 	SetColor(colorDefault);
 	for (int i = 0; i < 90; i++) { GotoXY(i + 5, lineStart + 1); putchar(196); }
 	GotoXY(5, lineStart + 2); 
-	cout << "| STT    | So hieu         | Loai may bay                           | So day   | So dong |";
+	cout << "| STT    | SO HIEU         | LOAI MAY BAY                           | SO DAY   | SO DONG |";
 	for (int i = 0; i < 90; i++) { GotoXY(i + 5, lineStart + 3); putchar(196); }
 
 	// print data
@@ -219,35 +230,25 @@ void QuanLyMayBay::ShowList()
 int QuanLyMayBay::Add()
 {
 	Clrscr();
-	GotoXY(0, 0); SetColor(colorGreen); cout << "THEM MAY BAY";
+	GotoXY(0, 0); SetColor(colorHeader); cout << "THEM MAY BAY";
 	
 	MayBay * maybay = new MayBay();
 
-	// Nhap so hieu
-	ClrLine(2); GotoXY(5, 2); SetColor(colorWhite); cout << "* Nhap so hieu may bay: ";
-	while(1) {
-		ClrLine(3);
-		GotoXY(10, 3); SetColor(colorYellow);
-		cin.ignore(); getline(cin, maybay->SoHieuMayBay);
-
-		if (!isSoHieuHopLe(maybay->SoHieuMayBay)) {
-			SetColor(colorRed); cout << "ERROR: So hieu khong hop le!";
-		}
-		else {
-			LineStandardize(maybay->SoHieuMayBay);
-			break;
-		}
-	}
+	maybay->SoHieuMayBay = Build_SoHieuMayBay();
+	ClrLine(2); GotoXY(5, 2); SetColor(colorWhite); cout << "So hieu may bay: ";
+	GotoXY(10, 3); SetColor(colorYellow); cout << maybay->SoHieuMayBay;
 
 	// Nhap loai
 	ClrLine(4); GotoXY(5, 4); SetColor(colorWhite); cout << "* Nhap loai may bay: ";
 	while (1) {
 		ClrLine(5);
 		GotoXY(10, 5); SetColor(colorYellow);
-		getline(cin, maybay->LoaiMayBay);
+		
+		if (input->GetInput() == -1) return -1;
+		maybay->LoaiMayBay = input->GetResult();
 
 		if (!isLoaiHopLe(maybay->LoaiMayBay)) {
-			SetColor(colorRed); cout << "ERROR: Loai khong hop le!";
+			ClrLine(6); GotoXY(0, 6); SetColor(colorRed); cout << "ERROR: Loai khong hop le!";
 		}
 		else {
 			LineStandardize(maybay->LoaiMayBay);
@@ -258,16 +259,16 @@ int QuanLyMayBay::Add()
 	// Nhap so day
 	ClrLine(6); GotoXY(5, 6); SetColor(colorWhite); cout << "* Nhap so day: ";
 	while (1) {
-		string soday;
 		ClrLine(7);
 		GotoXY(10, 7); SetColor(colorYellow);
-		getline(cin, soday);
+		
+		if (input->GetInput() == -1) return -1;
 
-		if (!isSoDayHopLe(soday)) {
-			SetColor(colorRed); cout << "ERROR: So day khong hop le!";
+		if (!input->isInteger()) {
+			ClrLine(8); GotoXY(0, 8); SetColor(colorRed); cout << "ERROR: So day khong hop le!";
 		}
 		else {
-			maybay->SoDay = StringToInteger(soday);
+			maybay->SoDay = StringToInteger(input->GetResult());
 			break;
 		}
 	}
@@ -275,16 +276,16 @@ int QuanLyMayBay::Add()
 	// Nhap so dong
 	ClrLine(8); GotoXY(5, 8); SetColor(colorWhite); cout << "* Nhap so dong: ";
 	while (1) {
-		string sodong;
 		ClrLine(9);
 		GotoXY(10, 9); SetColor(colorYellow);
-		getline(cin, sodong);
 
-		if (!isSoDongHopLe(sodong)) {
-			SetColor(colorRed); cout << "ERROR: So dong khong hop le!";
+		if (input->GetInput() == -1) return -1;
+
+		if (!input->isInteger()) {
+			ClrLine(10); GotoXY(0, 10); SetColor(colorRed); cout << "ERROR: So dong khong hop le!";
 		}
 		else {
-			maybay->SoDong = StringToInteger(sodong);
+			maybay->SoDong = StringToInteger(input->GetResult());
 			break;
 		}
 	}
@@ -298,20 +299,28 @@ int QuanLyMayBay::Add()
 	while (1) {
 		SetCursorVisible(1);
 		ClrLine(13); GotoXY(0, 13); SetColor(colorWhite); cout << "-> ";
-		cin >> choose;
+		
+		if (input->GetInput() == -1) return -1;
 
-		switch (choose)
-		{
-		case 1:
-			add(maybay);
-			GotoXY(0, 14); SetColor(colorCyan); cout << "INFO: Them thanh cong!";
-			Sleep(2000);
-			return -1;
-		case 2:
-			return -1;
-		default:
-			SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
-			break;
+		if (!input->isInteger()) {
+			ClrLine(14); GotoXY(0, 14); SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
+		}
+		else {
+			choose = StringToInteger(input->GetResult());
+
+			switch (choose)
+			{
+			case 1:
+				add(maybay);
+				ClrLine(14); GotoXY(0, 14); SetColor(colorCyan); cout << "INFO: Them thanh cong!";
+				Sleep(2000);
+				return -1;
+			case 2:
+				return -1;
+			default:
+				ClrLine(14); GotoXY(0, 14); SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
+				break;
+			}
 		}
 	}
 }
@@ -322,145 +331,161 @@ int QuanLyMayBay::Modify()
 
 	while (1) {
 		ClrLine(9); GotoXY(0, 9); SetColor(colorWhite); cout << "Nhap so thu tu trong danh sach: ";
-		SetColor(colorYellow); cin >> choose;
+		SetColor(colorYellow);
 
-		if (choose < 1 || choose > SoLuongMayBay) {
-			SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
+		if (input->GetInput() == -1) return -1;
+		if (!input->isInteger()) {
+			ClrLine(10); GotoXY(0, 10); SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
 		}
 		else {
-			Clrscr();
-			GotoXY(0, 0); SetColor(colorGreen); cout << "HIEU CHINH MAY BAY";
-			int index = choose - 1;
+			choose = StringToInteger(input->GetResult());
 
-			MayBay * maybay = new MayBay();
-			maybay->SoHieuMayBay = DanhSach[index]->SoHieuMayBay;
-			maybay->LoaiMayBay = DanhSach[index]->LoaiMayBay;
-			maybay->SoDay = DanhSach[index]->SoDay;
-			maybay->SoDong = DanhSach[index]->SoDong;
+			if (choose < 1 || choose > SoLuongMayBay) {
+				ClrLine(10); GotoXY(0, 10); SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
+			}
+			else {
+				Clrscr();
+				GotoXY(0, 0); SetColor(colorHeader); cout << "HIEU CHINH MAY BAY";
+				int index = choose - 1;
 
-			ClrLine(2); GotoXY(5, 2); SetColor(colorWhite); cout << "So hieu may bay: " << maybay->SoHieuMayBay;
+				MayBay * maybay = new MayBay();
+				maybay->SoHieuMayBay = DanhSach[index]->SoHieuMayBay;
+				maybay->LoaiMayBay = DanhSach[index]->LoaiMayBay;
+				maybay->SoDay = DanhSach[index]->SoDay;
+				maybay->SoDong = DanhSach[index]->SoDong;
 
-			// Nhap loai
-			ClrLine(4); GotoXY(5, 4); SetColor(colorWhite); cout << "Loai may bay: " << maybay->LoaiMayBay;
-			ClrLine(5); GotoXY(5, 5); SetColor(colorCyan); cout << "Hieu chinh Loai may bay? (Y/N) ";
+				ClrLine(2); GotoXY(5, 2); SetColor(colorWhite); cout << "So hieu may bay: " << maybay->SoHieuMayBay;
 
-			while (1) {
-				char c = InputKey();
-				bool ok = 0;
+				// Nhap loai
+				ClrLine(4); GotoXY(5, 4); SetColor(colorWhite); cout << "Loai may bay: " << maybay->LoaiMayBay;
+				ClrLine(5); GotoXY(5, 5); SetColor(colorCyan); cout << "Hieu chinh Loai may bay? (Y/N) ";
 
-				if (c == 'Y' || c == 'y') {
-					while (1) {
-						ClrLine(5); GotoXY(5, 5); SetColor(colorWhite); cout << "* Nhap loai may bay: ";
-						SetColor(colorYellow); getline(cin, maybay->LoaiMayBay);
+				while (1) {
+					char c = InputKey();
+					bool ok = 0;
 
-						if (!isLoaiHopLe(maybay->LoaiMayBay)) {
-							SetColor(colorRed); cout << "ERROR: Loai khong hop le!";
-						}
-						else {
-							LineStandardize(maybay->LoaiMayBay);
-							ok = 1;
-							break;
+					if (c == 'Y' || c == 'y') {
+						while (1) {
+							ClrLine(5); GotoXY(5, 5); SetColor(colorWhite); cout << "* Nhap loai may bay: ";
+							SetColor(colorYellow); 
+
+							if (input->GetInput() == -1) return -1;
+
+							if (!isLoaiHopLe(input->GetResult())) {
+								ClrLine(6); GotoXY(0, 6); SetColor(colorRed); cout << "ERROR: Loai khong hop le!";
+							}
+							else {
+								maybay->LoaiMayBay = LineStandardize(input->GetResult());
+								ok = 1;
+								break;
+							}
 						}
 					}
+					else if (c == 'N' || c == 'n') {
+						ClrLine(5);
+						cin.ignore();
+						break;
+					}
+
+					if (ok) break;
 				}
-				else if (c == 'N' || c == 'n') {
-					ClrLine(5);
-					cin.ignore();
-					break;
+
+				// Nhap so day
+				ClrLine(6); GotoXY(5, 6); SetColor(colorWhite); cout << "So day: " << maybay->SoDay;
+				ClrLine(7); GotoXY(5, 7); SetColor(colorCyan); cout << "Hieu chinh So day? (Y/N) ";
+
+				while (1) {
+					char c = InputKey();
+					bool ok = 0;
+
+					if (c == 'Y' || c == 'y') {
+						while (1) {
+							ClrLine(7); GotoXY(5, 7); SetColor(colorWhite); cout << "* Nhap so day: ";
+							SetColor(colorYellow);
+
+							if (input->GetInput() == -1) return -1;
+
+							if (!input->isInteger()) {
+								ClrLine(8); GotoXY(0, 8); SetColor(colorRed); cout << "ERROR: So day khong hop le!";
+							}
+							else {
+								maybay->SoDay = StringToInteger(input->GetResult());
+								ok = 1;
+								break;
+							}
+						}
+					}
+					else if (c == 'N' || c == 'n') {
+						ClrLine(7);
+						//cin.ignore();
+						break;
+					}
+
+					if (ok) break;
 				}
+
+				// Nhap so dong
+				ClrLine(8); GotoXY(5, 8); SetColor(colorWhite); cout << "So dong: " << maybay->SoDong;
+				ClrLine(9); GotoXY(5, 9); SetColor(colorCyan); cout << "Hieu chinh So dong? (Y/N) ";
+
+				while (1) {
+					char c = InputKey();
+					bool ok = 0;
+
+					if (c == 'Y' || c == 'y') {
+						while (1) {
+							ClrLine(9); GotoXY(5, 9); SetColor(colorWhite); cout << "* Nhap so dong: ";
+							SetColor(colorYellow); 
+							if (input->GetInput() == -1) return -1;
+
+							if (!input->isInteger()) {
+								ClrLine(10); GotoXY(0, 10); SetColor(colorRed); cout << "ERROR: So dong khong hop le!";
+							}
+							else {
+								maybay->SoDong = StringToInteger(input->GetResult());
+								ok = 1;
+								break;
+							}
+						}
+					}
+					else if (c == 'N' || c == 'n') {
+						ClrLine(9);
+						//cin.ignore();
+						break;
+					}
+
+					if (ok) break;
+				}
+
+				ClrLine(10);
+				GotoXY(5, 11); SetColor(colorCyan); cout << "1. Luu";
+				GotoXY(25, 11); SetColor(colorCyan); cout << "2. Tro ve";
+
+				while (1) {
+					SetCursorVisible(1);
+					ClrLine(13); GotoXY(0, 13); SetColor(colorWhite); cout << "-> ";
 					
-				if (ok) break;
-			}
+					if (input->GetInput() == -1) return -1;
+					if (!input->isInteger()) {
+						ClrLine(14); GotoXY(0, 14); SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
+					}
+					else {
+						choose = StringToInteger(input->GetResult());
 
-			// Nhap so day
-			ClrLine(6); GotoXY(5, 6); SetColor(colorWhite); cout << "So day: " << maybay->SoDay;
-			ClrLine(7); GotoXY(5, 7); SetColor(colorCyan); cout << "Hieu chinh So day? (Y/N) ";
-
-			while (1) {
-				char c = InputKey();
-				bool ok = 0;
-
-				if (c == 'Y' || c == 'y') {
-					while (1) {
-						string soday;
-
-						ClrLine(7); GotoXY(5, 7); SetColor(colorWhite); cout << "* Nhap so day: ";
-						SetColor(colorYellow); getline(cin, soday);
-
-						if (!isSoDayHopLe(soday)) {
-							SetColor(colorRed); cout << "ERROR: So day khong hop le!";
-						}
-						else {
-							maybay->SoDay = StringToInteger(soday);
-							ok = 1;
+						switch (choose)
+						{
+						case 1:
+							update_byposition(index, maybay);
+							ClrLine(14); GotoXY(0, 14); SetColor(colorCyan); cout << "INFO: Hieu chinh thanh cong!";
+							Sleep(2000);
+							return -1;
+						case 2:
+							return -1;
+						default:
+							ClrLine(14); GotoXY(0, 14); SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
 							break;
 						}
 					}
-				}
-				else if (c == 'N' || c == 'n') {
-					ClrLine(7);
-					//cin.ignore();
-					break;
-				}
-
-				if (ok) break;
-			}
-
-			// Nhap so dong
-			ClrLine(8); GotoXY(5, 8); SetColor(colorWhite); cout << "So dong: " << maybay->SoDong;
-			ClrLine(9); GotoXY(5, 9); SetColor(colorCyan); cout << "Hieu chinh So dong? (Y/N) ";
-
-			while (1) {
-				char c = InputKey();
-				bool ok = 0;
-
-				if (c == 'Y' || c == 'y') {
-					while (1) {
-						string sodong;
-
-						ClrLine(9); GotoXY(5, 9); SetColor(colorWhite); cout << "* Nhap so dong: ";
-						SetColor(colorYellow); getline(cin, sodong);
-
-						if (!isSoDongHopLe(sodong)) {
-							SetColor(colorRed); cout << "ERROR: So dong khong hop le!";
-						}
-						else {
-							maybay->SoDong = StringToInteger(sodong);
-							ok = 1;
-							break;
-						}
-					}
-				}
-				else if (c == 'N' || c == 'n') {
-					ClrLine(9);
-					//cin.ignore();
-					break;
-				}
-
-				if (ok) break;
-			}
-
-			ClrLine(10);
-			GotoXY(5, 11); SetColor(colorCyan); cout << "1. Luu";
-			GotoXY(25, 11); SetColor(colorCyan); cout << "2. Tro ve";
-
-			while (1) {
-				SetCursorVisible(1);
-				ClrLine(13); GotoXY(0, 13); SetColor(colorWhite); cout << "-> ";
-				cin >> choose;
-
-				switch (choose)
-				{
-				case 1:
-					update_byposition(index, maybay);
-					GotoXY(0, 14); SetColor(colorCyan); cout << "INFO: Hieu chinh thanh cong!";
-					Sleep(2000);
-					return -1;
-				case 2:
-					return -1;
-				default:
-					SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
-					break;
 				}
 			}
 		}
@@ -473,27 +498,36 @@ int QuanLyMayBay::Delete()
 
 	while (1) {
 		ClrLine(9); GotoXY(0, 9); SetColor(colorWhite); cout << "Nhap so thu tu trong danh sach: ";
-		SetColor(colorYellow); cin >> choose;
+		SetColor(colorYellow); 
+		
+		if (input->GetInput() == -1) return -1;
 
-		if (choose < 1 || choose > SoLuongMayBay) {
-			SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
+		if (!input->isInteger()) {
+			ClrLine(10); GotoXY(0, 10); SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
 		}
 		else {
-			int index = choose - 1;
-			ClrLine(10); GotoXY(0, 10); SetColor(colorCyan); cout << "Ban co chac chan xoa? (Y/N) ";
+			choose = StringToInteger(input->GetResult());
 
-			while (1) {
-				char c = InputKey();
+			if (choose < 1 || choose > SoLuongMayBay) {
+				ClrLine(10); GotoXY(0, 10); SetColor(colorRed); cout << "ERROR: Input khong hop le!" << endl;
+			}
+			else {
+				int index = choose - 1;
+				ClrLine(10); GotoXY(0, 10); SetColor(colorCyan); cout << "Ban co chac chan xoa? (Y/N) ";
 
-				if (c == 'Y' || c == 'y') {
-					delete_byposition(index);
-					ClrLine(10); GotoXY(0, 10); SetColor(colorCyan); cout << "INFO: Xoa thanh cong!";
-					Sleep(2000);
-					return -1;
-				}
-				else if (c == 'N' || c == 'n') {
-					cin.ignore();
-					return -1;
+				while (1) {
+					char c = InputKey();
+
+					if (c == 'Y' || c == 'y') {
+						delete_byposition(index);
+						ClrLine(10); GotoXY(0, 10); SetColor(colorCyan); cout << "INFO: Xoa thanh cong!";
+						Sleep(2000);
+						return -1;
+					}
+					else if (c == 'N' || c == 'n') {
+						cin.ignore();
+						return -1;
+					}
 				}
 			}
 		}
